requirements.txt
Pillow==5.3.0
opencv-python==3.4.5
graphviz=0.10.1